+++ src/history_book/api/main.py
@@ -3,7 +3,7 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-from .routes import agent, books, chat, kg, kg_metrics
+from .routes import books, chat, kg, kg_metrics
 
 
 def create_app() -> FastAPI:
@@ -25,7 +25,6 @@ def create_app() -> FastAPI:
 
     # Include routers
     app.include_router(chat.router, prefix="/api")
-    app.include_router(agent.router, prefix="/api")
     app.include_router(books.router, prefix="/api")
     app.include_router(kg.router, prefix="/api")
     app.include_router(kg_metrics.router, prefix="/api")
